refactor: rename utils to lib
