+++ src/2022/1/run.ts
@@ -1,5 +1,5 @@
-import { getInput } from '../../utils/getInput';
-import { log } from '../../utils/log';
+import { getInput } from '../../lib/getInput';
+import { log } from '../../lib/log';
 import { collectCaloriesTotals } from './collectCaloriesTotals';
 import { getHighest, getHighestThree } from './getHighestCalories';
