+++ src/lib/getInput.ts
