+++ src/lib/isTruthy.ts
